Make cmath.rs a single file
